pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "devassist"
version = "0.1.0"
description = "AI agent for developer productivity built with Pydantic AI"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai[openai]",
    "httpx",
    "python-dotenv",
    "fastmcp",
    "rich",
]

[project.scripts]
devassist = "devassist.cli:main"